+++ djmoney/contrib/exchange/migrations/0001_initial.py
@@ -3,6 +3,8 @@
 import django.db.models.deletion
 from django.db import migrations, models
 
+from djmoney.settings import CURRENCY_CODE_MAX_LENGTH
+
 
 class Migration(migrations.Migration):
 
@@ -16,14 +18,14 @@ class Migration(migrations.Migration):
             fields=[
                 ("name", models.CharField(max_length=255, primary_key=True, serialize=False)),
                 ("last_update", models.DateTimeField(auto_now=True)),
-                ("base_currency", models.CharField(max_length=3)),
+                ("base_currency", models.CharField(max_length=CURRENCY_CODE_MAX_LENGTH)),
             ],
         ),
         migrations.CreateModel(
             name="Rate",
             fields=[
                 ("id", models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name="ID")),
-                ("currency", models.CharField(max_length=3)),
+                ("currency", models.CharField(max_length=CURRENCY_CODE_MAX_LENGTH)),
                 ("value", models.DecimalField(decimal_places=6, max_digits=20)),
                 (
                     "backend",