+++ djmoney/settings.py
@@ -1,4 +1,6 @@
 import operator
+import warnings
+from types import MappingProxyType
 
 from django.conf import settings
 
@@ -23,9 +25,14 @@ if CURRENCY_CHOICES is None:
 
 CURRENCY_CHOICES.sort(key=operator.itemgetter(1, 0))
 DECIMAL_PLACES = getattr(settings, "CURRENCY_DECIMAL_PLACES", 2)
-DECIMAL_PLACES_DISPLAY = getattr(
-    settings, "CURRENCY_DECIMAL_PLACES_DISPLAY", {currency[0]: DECIMAL_PLACES for currency in CURRENCY_CHOICES}
-)
+_decimal_display_value = getattr(settings, "CURRENCY_DECIMAL_PLACES_DISPLAY", None)
+if _decimal_display_value is None:
+    warnings.warn(
+        "`CURRENCY_DECIMAL_PLACES_DISPLAY` is deprecated and will be removed in django-money 3.0.",
+        DeprecationWarning,
+    )
+DECIMAL_PLACES_DISPLAY = _decimal_display_value or {currency[0]: DECIMAL_PLACES for currency in CURRENCY_CHOICES}
+IS_DECIMAL_PLACES_DISPLAY_SET = _decimal_display_value is not None
 
 OPEN_EXCHANGE_RATES_URL = getattr(settings, "OPEN_EXCHANGE_RATES_URL", "https://openexchangerates.org/api/latest.json")
 OPEN_EXCHANGE_RATES_APP_ID = getattr(settings, "OPEN_EXCHANGE_RATES_APP_ID", None)
@@ -34,3 +41,7 @@ FIXER_ACCESS_KEY = getattr(settings, "FIXER_ACCESS_KEY", None)
 BASE_CURRENCY = getattr(settings, "BASE_CURRENCY", "USD")
 EXCHANGE_BACKEND = getattr(settings, "EXCHANGE_BACKEND", "djmoney.contrib.exchange.backends.OpenExchangeRatesBackend")
 RATES_CACHE_TIMEOUT = getattr(settings, "RATES_CACHE_TIMEOUT", 600)
+
+CURRENCY_CODE_MAX_LENGTH = getattr(settings, "CURRENCY_CODE_MAX_LENGTH", 3)
+
+MONEY_FORMAT = MappingProxyType(getattr(settings, "MONEY_FORMAT", {}))